refactor(rig): group agents by modality, drop tool plumbing, README rewrite
Three connected moves that touch the same surface:

## Agent layout grouped by modality

`agent/cv/`  bundles `classify` (was top-level `cv`) and `verify`
(was `cv_verify_agent`).  `agent/ner/` bundles `detect` (was
top-level `ner`) and `verify` (was `ner_verify_agent`). The
classify/verify/detect submodules are private; each parent re-
exports its types so external paths flatten one level:

    nvisy_rig::agent::cv::CvAgent
    nvisy_rig::agent::cv::CvVerifyAgent
    nvisy_rig::agent::ner::NerAgent
    nvisy_rig::agent::ner::NerVerifyAgent

`generate` stays top-level (single agent, no detect/verify split).
`agent::*` itself no longer re-exports concrete agents; only the
cross-cutting types (`AgentConfig`, `AgentProvider`, `DetectionConfig`,
`UsageStats`) sit at that level. Old flat `nvisy_rig::agent::CvAgent`
paths no longer work; consumers (pipelines, engine) updated.

Source moved via `git mv` so history follows the files.

## Tool plumbing deleted (#158)

`BaseAgentBuilder::tool` + the `tools: Vec<Box<dyn ToolDyn>>` field
+ the typestate fork in `build_rig_agent` were carrying a
`-D warnings` dead-code failure with no current caller. The
original tool-using design (`CvProvider` + `CvRigTool`) was removed
earlier when CV switched to pre-computed detection inputs.

Issue #158 captures the design thinking for a future return — most
likely shape is "CV agent calls OCR as a tool" with the OCR adapter
living either in nvisy-ocr or in nvisy-rig::agent::cv behind a
feature flag.

## README rewrite

`nvisy-rig/README.md` rewritten to match the nvisy-pattern README
shape: badge → tagline → `## Overview` with prose explaining the
modality grouping + pipeline surface + a minimal NerPipeline code
example → `## Documentation` / `## Changelog` / `## License` /
`## Support` boilerplate. Stale references (`agent::base` is
pub(crate), `agent::ocr` doesn't exist, `nvisy-detection` was
disbanded) all replaced with accurate types.

## Plus

- Fixed a stale `useless_conversion` clippy in
  `nvisy-engine/src/operation/detection.rs` (leftover
  `.map_err(nvisy_core::Error::from)?` from when the engine
  returned a detection-local error).
- 13 broken intra-doc links across rig + engine fixed:
  `crate::agent::FooAgent` paths re-pointed to modality-qualified
  `crate::agent::cv::FooAgent` / `crate::agent::ner::FooAgent`;
  stale `super::BaseAgent` and `[classify]`/`[detect]`/`[verify]`
  links to now-private submodules dropped.
- Five inline `[Label](url)` rustdoc links in module docstrings
  (BCP-47, Luhn algorithm, RFC 8594, Deprecation header draft,
  GitHub issue #151) converted to reference-style.
