+++ crates/nvisy-ontology/src/provenance/redaction_map.rs
@@ -7,10 +7,11 @@
 //!
 //! A future extension may pair this index with a separate blob store
 //! to support reversibility for image/audio modalities: see
-//! [issue #151](https://github.com/nvisycom/runtime/issues/151).
+//! [issue #151].
 //!
 //! [`AuditEntry::value`]: super::AuditEntry::value
 //! [`RedactionValue`]: super::RedactionValue
+//! [issue #151]: https://github.com/nvisycom/runtime/issues/151
 
 use derive_more::{Deref, DerefMut};
 use schemars::JsonSchema;
