+++ crates/nvisy-rig/src/agent/cv/classify/output.rs
