+++ crates/nvisy-rig/src/agent/cv/classify/prompt.rs
