+++ crates/nvisy-rig/src/agent/cv/verify/input.rs
