+++ crates/nvisy-rig/src/agent/cv/verify/prompt.rs
