+++ crates/nvisy-rig/src/agent/mod.rs
@@ -1,25 +1,21 @@
-//! Specialized LLM agents: NER (text), CV (vision), generation
-//! (synthetic replacement values), and entity verification
-//! (LLM-mediated check of upstream-proposed entities against an
-//! image).
+//! Specialized LLM agents, grouped by modality.
 //!
-//! Each agent composes a `BaseAgent` with domain-specific prompts
-//! and optional tools. Public types are re-exported from [`crate`]
-//! — consumer code should not reach into submodules.
+//! Each modality (CV, NER) bundles its detect-style and
+//! verify-style agents under a shared parent — they share data
+//! shapes ([`cv::CvEntity`], [`ner::NerCandidate`])
+//! and are meant to be used together by orchestrating pipelines.
+//!
+//! Cross-cutting infrastructure ([`AgentConfig`], [`AgentProvider`],
+//! [`DetectionConfig`], [`UsageStats`]) is re-exported at this
+//! level since every agent consumes it.
+
+pub mod cv;
+pub mod generate;
+pub mod ner;
 
 mod base;
-mod cv;
-mod cv_verify_agent;
-mod generate;
-mod ner;
-mod ner_verify_agent;
 
 #[cfg(any(feature = "openai-whisper", feature = "openai-tts"))]
 pub(crate) use self::base::AuthenticatedProvider;
-pub(crate) use self::base::{ALL_TYPES_HINT, BaseAgent, UnauthenticatedProvider};
+pub(crate) use self::base::{ALL_TYPES_HINT, UnauthenticatedProvider};
 pub use self::base::{AgentConfig, AgentProvider, DetectionConfig, UsageStats};
-pub use self::cv::{CvAgent, CvDetection, CvEntity};
-pub use self::cv_verify_agent::{CvVerifyAgent, VerificationCandidate};
-pub use self::generate::{GenAgent, GenRequest, GeneratedEntity};
-pub use self::ner::{KnownNerEntity, NerAgent, NerCandidate, NerContext};
-pub use self::ner_verify_agent::{NerVerifyAgent, UnresolvedCandidatePolicy};
