+++ crates/nvisy-rig/src/agent/ner/detect/context.rs
