+++ crates/nvisy-rig/src/agent/ner/detect/prompt.rs
