+++ crates/nvisy-rig/src/agent/ner/verify/localize.rs
